refactor(update-weight): dissolve common.py
The TP/ETP gather machinery (all_gather_params_async and helpers) moves into
its only consumer, the direct iterator; named_params_and_buffers becomes
megatron_utils/named_weights.py — model introspection shared by the actor,
hf_export, and the iterator, never update-specific.
