+++ miles/backends/megatron_utils/named_weights.py
@@ -1,0 +1,130 @@
+"""Enumerate a megatron model's named parameters and buffers.
+
+Global naming folds virtual-PP and EP offsets into layer indices
+(decoder.layers, mtp.layers, expert_bias) so names are comparable across
+ranks; witness params are skipped.
+"""
+
+import inspect
+import re
+from argparse import Namespace
+from collections.abc import Iterator, Sequence
+
+import torch
+from megatron.core.transformer.transformer_layer import get_transformer_layer_offset
+
+from miles.backends.megatron_utils.misc_utils import strip_param_name_prefix
+from miles.backends.training_utils.parallel import get_parallel_state
+
+
+def named_params_and_buffers(
+    args: Namespace,
+    model: Sequence[torch.nn.Module],
+    convert_to_global_name: bool = True,
+) -> Iterator[tuple[str, torch.Tensor]]:
+    if convert_to_global_name:
+        return _named_params_and_buffers_global(args, model)
+    return _named_params_and_buffers_vanilla(model)
+
+
+def _named_params_and_buffers_vanilla(model: Sequence[torch.nn.Module]) -> Iterator[tuple[str, torch.Tensor]]:
+    for vp_stage, model_module in enumerate(model):
+
+        def _compute_fqn(name, vp_stage=vp_stage):
+            return f"vp_stages.{vp_stage}.{strip_param_name_prefix(name)}"
+
+        for name, param in model_module.named_parameters():
+            if getattr(param, "_is_witness_param", False):
+                continue
+            yield _compute_fqn(name), param
+
+        for name, buffer in model_module.named_buffers():
+            # TODO shall we handle (almost) all buffers like Megatron Bridge
+            if "expert_bias" not in name:
+                continue
+            yield _compute_fqn(name), buffer
+
+
+def _named_params_and_buffers_global(
+    args: Namespace, model: Sequence[torch.nn.Module]
+) -> Iterator[tuple[str, torch.Tensor]]:
+    """
+    Yield (global_name, param/buffer) with consistent names across PP/EP. Adjusts indices for
+    virtual PP + EP offsets. Handles decoder.layers, mtp.layers (Multi-Token Prediction), expert_bias.
+    """
+    ep_size = get_parallel_state().ep.size
+    ep_rank = get_parallel_state().ep.rank
+    if args.num_experts:
+        expert_offset = ep_rank * args.num_experts // ep_size
+
+    sig = inspect.signature(get_transformer_layer_offset)
+    need_vp_stage = "vp_stage" in sig.parameters
+
+    for vp_stage, model_module in enumerate(model):
+        if need_vp_stage:
+            layer_offset = get_transformer_layer_offset(model_module.config, vp_stage)
+        else:
+            layer_offset = get_transformer_layer_offset(model_module.config)
+        for name, param in model_module.named_parameters():
+            if getattr(param, "_is_witness_param", False):
+                continue
+            # for model without ddp wrap
+            if not name.startswith("module.module."):
+                name = "module." + name
+
+            decoder_layers_pattern = r"module\.module\.(?:language_model\.)?decoder\.layers\.(\d+)\.(.+)"
+            match = re.match(decoder_layers_pattern, name)
+            if not match:
+                # MTP (Multi-Token Prediction) layers for speculative decoding
+                mtp_layers_pattern = r"module\.module\.mtp\.layers\.(\d+)\.(.+)"
+                match = re.match(mtp_layers_pattern, name)
+                if not match:
+                    yield name, param
+                    continue
+
+                # MTP layer indices start from 0
+                layer_idx, rest = match.groups()
+                expert_pattern = r"(transformer_layer|mtp_model_layer)\.mlp\.experts\.(.+)\.weight(\d+)"
+                match = re.match(expert_pattern, rest)
+                if not match:
+                    yield name, param
+                    continue
+
+                inner, rest, expert_idx = match.groups()
+                expert_idx = int(expert_idx) + expert_offset
+                yield (
+                    f"module.module.mtp.layers.{layer_idx}.{inner}.mlp.experts.{rest}.weight{expert_idx}",
+                    param,
+                )
+                continue
+
+            layer_idx, rest = match.groups()
+            layer_idx = int(layer_idx) + layer_offset
+
+            # this is hardcoded for te grouped matmul
+            expert_pattern = r"mlp.experts\.(.+)\.weight(\d+)"
+            match = re.match(expert_pattern, rest)
+            if match:
+                rest, expert_idx = match.groups()
+                expert_idx = int(expert_idx) + expert_offset
+                yield f"module.module.decoder.layers.{layer_idx}.mlp.experts.{rest}.weight{expert_idx}", param
+            else:
+                yield f"module.module.decoder.layers.{layer_idx}.{rest}", param
+
+        # treat expert bias as normal parameters
+        for name, buffer in model_module.named_buffers():
+            # TODO shall we handle (almost) all buffers like Megatron Bridge
+            if "expert_bias" not in name:
+                continue
+            # for model without ddp wrap
+            if not name.startswith("module.module."):
+                name = "module." + name
+
+            decoder_layers_pattern = r"module\.module\.(?:language_model\.)?decoder\.layers\.(\d+)\.(.+)"
+            match = re.match(decoder_layers_pattern, name)
+            if not match:
+                yield name, buffer
+            else:
+                layer_idx, rest = match.groups()
+                layer_idx = int(layer_idx) + layer_offset
+                yield f"module.module.decoder.layers.{layer_idx}.{rest}", buffer
